octofit-tracker/backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
pymongo>=4.0,<5.0
python-decouple>=3.6,<4.0
django-cors-headers>=3.5,<4.0